Added more code for dungeon generation
Since I have not designed a overall framework, things are getting
frustratingly messy.  I didn't want to have to design some complex
system for this, and I was hoping I could keep it simpler, but I
need a minimum set of features that just ends up being a real
hassle.