Fixed spawning, level generation, detail tiles, can exit dungeon
Only the logic for the stack dungeon type is done for now.  Going
up and down floors works great.  I somehow seemed to make every
stupid mistake I could when spawning the player.  I got all sorts
of things backwards.  This is probably a disadvantage of working
on this only when it's late....