Added transition animations between modes
Now when entering a new area, there's a nice quick transition.
When leaving the previous mode, the screen contents are frozen
by enabling screen persistence until the transition is halfway
complete and the previous screen contents are covered.  Then the
next mode starts to step and draw, showing the new contents
while finishing the transition.  This also means that the player
can resume interacting with the game as soon as the new mode starts
stepping.