Added dialogue prototype
For now it simply displays a static line of text if the player
walks left as a quick test.  It determines when a set of dialogue
is done when it no longer has newlines at the start of subsequent
lines.

In order to actually display text properly, I'll need to add the
infrastructure for saving the game, at least partially, because I
will need status variables to determine whether a specific piece of
dialogue has happened yet or not.