Added preliminary boss dialogue for the catpaw and boss encounters
Right now, you can run from boss encounters, which is undesirable.
Also, the code for all this scripted dialogue is very hacky,
because I feel like any framework for it would be too big.  I'm
going to keep all the dialogue and boss events hard-coded for now.

The code size is at 99% of progmem, so I may end up removing a few
dungeon themes, since some of them are so similar.  I probably
should also remove the top-of-wall texture from the themes and make
it a common tile.  Of course, I will need to do other optimizations
as well if I want to fit everything without trying to resort to
using the SD card.