Merged several themes, removed two dungeons, removed one enemy type
Some themes were somewhat redundant, and some dungeons were behind
areas that are currently closed off.  I was planning to open up the
bigger warehouse area, but now I think I'll leave it closed and not
even keep those dungeons.  With the removal of the warehouse, I am
also removing the OVERMAN enemy type, since he was supposed to be
the boss.