Compressed world map with run-length encoding
I compressed the map with very simple run-length encoding, in
which there are basically a bunch of (count,type) pairs.  Doing
this saved 1072 bytes, even including the decompression code.  The
size went from 30358 (99%) to 29284 (95%).  I have ideas for
further compression I might be able to do, but I'm not sure it will
provide enough benefit for the increase in decompression bytes.

The world_get function now gets requested map indices, but it is
slow to retrieve an index near the end of the map, since we must
linearly search forward for the index.  I alleviate this by keeping
a global index so that the most expensive and frequent source of
reading the map (drawing the map tiles) incrementally searches
through the map data instead of starting from the beginning every
time world_get is invoked.  As long as a subsequent call to
world_get asks for an index higher than the previous call, the
search can start at the point where the previous index was found.