Removed several dungeons, fixed text bug
When I compressed the map, I also made changes to the map in which
I removed some dungeons to save a bit more space.  Now the data for
those dungeons has been removed as well.

I also fixed the indices for the text events, as they were off by
one due to my rewording of the script that removed a line.