Moved shadow dialogue trigger and spawn point, added intro text
The intro text really needs to move up and not have a space for
a speaker, since it's basically narrated.  Some of the text that is
displayeed goes onto 4 lines, so it will either need to be modified
or I need to have an option to display it one line up (perhaps if
the names list is null?).

Also, when I added the intro text, it started before the earliest
text I was displaying, which meant I started the first status to
be at -1.  This was a quick hack, and I think I want to design this
differently in the future.