Merged seek functionality into fill_dialogue_buffer
The separate dialogue_seek function was a waste of space, as it
and fill_dialogue_buffer shared a lot of identical code.  Merging
it into fill_dialogue_buffer saved another 128 bytes, taking game
size down to 29654 (96%), which I'm fairly happy with.  Hopefully
it will provide enough space for the rest of the dialogue and other
features.