Removed shadow from initial party, add later, tons of rebalancing
Now the shadow joins the party when he says he does.

I also did all sorts of rebalances.  I doubled the amount of health
of everyone (enemies and allies) but kept damage the same so that
you have more time to buff up.  I also increased the effects of
stat boosting abilities and items to make them more useful, and
altered the drop rates so that the rare items are more common.