Added draft of final game dialogue
The final dialogue increases the game size from 29772 (97%) to
30464 (99%), which is both concerning in that there is very little
space left, but also reassuring that it didn't immediately push
the game over 100%.  I do still need to actually add all the game
triggers for the battles and dialogue, so I can't say the dialogue
actually fits yet.  I think I will need to store the triggers in
such a way that I use less code for it, so I need to determine an
efficient way to store that data.