Organized game events to be more efficient
Before, I had an ad-hoc set of if statements to determine when to
display dialogue or have a boss battle.  Now, it is a data
structure accessed by a single function that loops through it and
determines if the conditions are valid for an event to occur.

Unfortunately, the benefits were very minor for now, as I only
shrunk it by 14 bytes and went to 30450 (99%).  However, as I add
more events, I think this will have increasing benefits.

I did not bother to create anything for events in the overworld
because I have so few of them, and I would need a lot of bytes to
describe the collision boxes I would need for certain events.  I
might even try to move one or two events into dungeons if I can.

As I was thinking of other ways to save space, I recalled that I
set the Gamebuino library to have 3 sound channels for my music
tracker project, so I went and changed the settings to have one
channel, remove grayscale support, and disable the extended note
range.  This reduced the game size all the way to 29440 (96%),
which is a fairly significant space gain.