Removed name_array field from dialogue_event struct, moved dialogue
I changed the thug boss to be a bruiser instead, and this meant
that the indices between the player_names and enemy_names were
easily distinguishable.  This allowed me to remove the field that
specified which name array the name belonged to, saving a byte per
entry in dialogue_events.  Unfortunately, this only reduced the
game size by 10 bytes to 29430 (96%).  This should still help when
I am adding more entries in the future.

I also decided to move one of the dialogue events on the overworld
to a dungeon, and this saved quite a bit more: 48 bytes, bringing
the size down to 29382 (96%).  Obviously having an if statement for
every dialogue event is quite a lot more costly than a data
structure, once you get over the initial cost.