Added all side quests
Now the game content is basically complete and I believe the game
can be played from start to finish, including all side quests.  I
haven't beaten the whole game from start to finish because I keep
accidentally triggering the "run bug" in which running in a dungeon
breaks the game, so I need to fix that.  It should be an easy fix.

The combat is still unbalanced and there's no saving yet.  Plus,
there's not a proper title screen, game over screen, or victory
screen, so there's a lot of balances and fixes that need to be
done.  That isn't to mention the changes I need to make to the
inventory system to add a maximum cap to the number you are allowed
to hold for each item, and handling for an empty inventory.

The side quests brought the game size to 29830 (97%) up from 29690
(97%).  I still have a bit of space, but I'm worried about the size
of the EEPROM library.  If that fits, I think there won't be too
many space problems, but if it doesn't, I might have to make some
changes.