Added saving and loading, changed main menu and pause menu
Now you can save and load the game as long as you are on the
overworld.  I also got rid of the default title screen, which saves
some space, and added a minimalist title screen that lets you
return to the loader and load a saved game.  I also changed the
pause menu to allow the game to be saved.  When not in the
overworld, the menu does not show the option to save because I did
not want to have to handle saving dungeon data, and it makes the
dungeons a bit more challenging, since the player will have to plan
ahead for each dungeon.