Added maximum inventory capacity, added handling of empty inventory
Now, if the player has more than 16 of an item, the mudlark uses
that item immediately after finding it instead of storing it. Also,
if the player doesn't have any food or drink, those menus won't
even do anything when the player tries to select them.

I also changed the point where the random number generator is
seeded to be after starting or loading a game.