Rebalanced enemy levels, decreased liquor effectiveness
Now bosses are much harder than other enemies encountered in their
area.  Also, I made it so that liquor can never reduce damage to
zero (minimum enemy damage is now 1), and it doesn't increase
damage resistance as much.

I also removed the bug that caused the game to glitch if the player
ran from a battle in a dungeon.