Changed dungeon theme, made minor changes and fix to overworld
The warehouse/brick theme was too similar to the catpaw/house
theme, so I changed it to have crates.

One section in the overworld had the wrong tiles (water that didn't
animate because I had put the second tile instead of the first), so
I fixed that and also reduced variation in areas out of range of
the player.  Those areas that are out of reach of the player can
omit details, and this saves more space due to the run-length
compression.