Added more enemy graphics, made small optimizations to fit them
The additional enemy images are a sailor, doctor, golem, and the
final boss himself, the leader.  Unfortunately, each image is 18
more bytes, and this pushed me over the size limit.

I did some changes to remove duplicated code in a couple places
by hoisting the code into a function that calls the two separate
functions using that code.  It was behavior that could be done
regardless of the game mode (incrementing animation frames).  I
also made some smaller tweaks that freed up a few bytes here and
there, and some cleanup changes that didn't really change much of
anything.