Fixed damage overflow bug, corrected game completion percentage
I changed the data type for damage to a 16-bit integer instead of
an 8-bit integer so it would not overflow when more than 255 damage
is done.

Also, the game completion percentage being displayed was half of
what it should have been because the calculation was incorrect.
Now it should correctly display what percentage of the game has
been completed.